Separated geminiapi into geminiapi and geminihandler
GeminiApi now only handles the individual api calls
GeminiHandler handles the calling of chunking/batching functions and then calls a function in GeminiApi to make the necessary API calls
These classes are likely to be renamed in the future for clarity.
