+++ gemini_efficient_api_calls/__init__.py
@@ -1,6 +1,7 @@
+from .geminihandler import GeminiHandler
 from .geminiapi import GeminiApi
-from .chunker import Chunker
 from .input_handler.textinputs import BaseTextInput, FileInput, WebsiteInput
 from .processor.textchunkandbatch import TextChunkAndBatch
+from .processor.mediachunkandbatch import MediaChunkAndBatch
 
 __version__ = "0.1.1"
