+++ gemini_efficient_api_calls/geminiapi.py
@@ -6,7 +6,6 @@ from typing import Any
 from google import genai
 from google.genai import types, errors
 
-from .chunker import Chunker
 from .mediachunker import MediaChunker
 
 from .input_handler.textinputs import BaseTextInput
@@ -41,10 +40,11 @@ class GeminiApi:
     def __init__(
         self,
         api_key : str,
-        model : str
+        model : str,
+        **kwargs
     ):
         # Error handling - Api key not correct
-        self.client = genai.Client(api_key=api_key)
+        self.client = genai.Client(api_key=api_key, **kwargs)
 
         # Error handling - Model not correct, default to a model
         self.model = model
@@ -60,7 +60,7 @@ class GeminiApi:
     def generate_content(
         self,
         prompt : str,
-        file = None,
+        files = [],
         system_prompt : str = None,
         max_retries : int = 5,
     ):
@@ -71,9 +71,13 @@ class GeminiApi:
         if system_prompt == None:
             system_prompt = DEFAULT_SYSTEM_PROMPT
 
+        if len(files) != 0:
+            prompt = [prompt]
+            for file in range(len(files)):
+                uploaded_file = self.client.files.upload(file=file)
+                prompt.append(uploaded_file)
+
         for i in range(max_retries):
-            if file:
-                prompt = [prompt, file]
             try:
                 response = self.client.models.generate_content(
                     model=self.model,
@@ -116,198 +120,198 @@ class GeminiApi:
             output_tokens = 0
         )
 
-    def generate_content_fixed(
-        self,
-        content : BaseTextInput,
-        questions : list[str],
-        chunk_char_length : int = 100000,
-        questions_per_batch : int = 50,
-        window_char_length : int = 100,
-        system_prompt : str = None
-    ):
-        # TODO: Currently can only handle a text response i.e. not a code block.
+    # def generate_content_fixed(
+    #     self,
+    #     content : BaseTextInput,
+    #     questions : list[str],
+    #     chunk_char_length : int = 100000,
+    #     questions_per_batch : int = 50,
+    #     window_char_length : int = 100,
+    #     system_prompt : str = None
+    # ):
+    #     # TODO: Currently can only handle a text response i.e. not a code block.
 
-        # # Chunking and Batching the questions
-        # chunker = Chunker()
-        # if enable_sliding_window:
-        #     chunks = chunker.sliding_window_chunking_by_size(content, chunk_char_length, window_char_length)
-        # else:
-        #     chunks = chunker.fixed_chunking_by_size(content, chunk_char_length)
-        # question_batches = chunker.fixed_question_batching(questions, questions_per_batch)
+    #     # # Chunking and Batching the questions
+    #     # chunker = Chunker()
+    #     # if enable_sliding_window:
+    #     #     chunks = chunker.sliding_window_chunking_by_size(content, chunk_char_length, window_char_length)
+    #     # else:
+    #     #     chunks = chunker.fixed_chunking_by_size(content, chunk_char_length)
+    #     # question_batches = chunker.fixed_question_batching(questions, questions_per_batch)
 
-        chunks = TextChunkAndBatch.chunk_sliding_window_by_length(
-            text_input = content,
-            chunk_char_size = chunk_char_length,
-            window_char_size = window_char_length
-        )
-        question_batches = TextChunkAndBatch.batch_by_number_of_questions(
-            questions = questions, 
-            questions_per_batch = questions_per_batch
-        )
+    #     chunks = TextChunkAndBatch.chunk_sliding_window_by_length(
+    #         text_input = content,
+    #         chunk_char_size = chunk_char_length,
+    #         window_char_size = window_char_length
+    #     )
+    #     question_batches = TextChunkAndBatch.batch_by_number_of_questions(
+    #         questions = questions, 
+    #         questions_per_batch = questions_per_batch
+    #     )
 
-        answers = {}
+    #     answers = {}
 
-        total_input_tokens = 0
-        total_output_tokens = 0
+    #     total_input_tokens = 0
+    #     total_output_tokens = 0
 
-        for batch in question_batches:
-            for chunk in chunks:
-                query_contents = f'Content:\n{chunk}\n\nThere are {len(batch)} questions. The questions are:\n' + '\n\t- '.join(batch)
-                response = self.generate_content(query_contents, system_prompt=system_prompt)
+    #     for batch in question_batches:
+    #         for chunk in chunks:
+    #             query_contents = f'Content:\n{chunk}\n\nThere are {len(batch)} questions. The questions are:\n' + '\n\t- '.join(batch)
+    #             response = self.generate_content(query_contents, system_prompt=system_prompt)
 
-                total_input_tokens += response.input_tokens
-                total_output_tokens += response.output_tokens
+    #             total_input_tokens += response.input_tokens
+    #             total_output_tokens += response.output_tokens
 
-                # TODO: If the question has already been answered in a previous chunk the new answer is disregarded, this can be
-                # further optimised so the question is not asked again.
-                for i in range(len(response.content)):
-                    if batch[i] not in answers.keys() and response.content[i] !=  'N/A':
-                        answers[batch[i]] = response.content[i]
+    #             # TODO: If the question has already been answered in a previous chunk the new answer is disregarded, this can be
+    #             # further optimised so the question is not asked again.
+    #             for i in range(len(response.content)):
+    #                 if batch[i] not in answers.keys() and response.content[i] !=  'N/A':
+    #                     answers[batch[i]] = response.content[i]
         
-        # TODO: Better way of returning? Tuple?
-        return Response(
-            content = answers,
-            input_tokens = total_input_tokens,
-            output_tokens = total_output_tokens
-        )
+    #     # TODO: Better way of returning? Tuple?
+    #     return Response(
+    #         content = answers,
+    #         input_tokens = total_input_tokens,
+    #         output_tokens = total_output_tokens
+    #     )
     
-    def generate_content_token_aware(
-        self,
-        content : BaseTextInput,
-        questions : list[str],
-        system_prompt : str = None
-    ):
-        # A version of generate_content_fixed() that automatically chunks depending on the token limits of the model being used.
+    # def generate_content_token_aware(
+    #     self,
+    #     content : BaseTextInput,
+    #     questions : list[str],
+    #     system_prompt : str = None
+    # ):
+    #     # A version of generate_content_fixed() that automatically chunks depending on the token limits of the model being used.
         
-        # Adding default system prompt if one is not given.
-        if system_prompt == None:
-            system_prompt = DEFAULT_SYSTEM_PROMPT
+    #     # Adding default system prompt if one is not given.
+    #     if system_prompt == None:
+    #         system_prompt = DEFAULT_SYSTEM_PROMPT
         
-        model_info = self.client.models.get(model=self.model)
-        input_token_limit = model_info.input_token_limit
-        output_token_limit = model_info.output_token_limit
+    #     model_info = self.client.models.get(model=self.model)
+    #     input_token_limit = model_info.input_token_limit
+    #     output_token_limit = model_info.output_token_limit
 
-        total_input_tokens = 0
-        total_output_tokens = 0
+    #     total_input_tokens = 0
+    #     total_output_tokens = 0
 
-        answers = {}
-        queue = [(content.content, questions)]
+    #     answers = {}
+    #     queue = [(content.content, questions)]
 
-        while len(queue) > 0:
-            curr_content, curr_questions = queue.pop(0)
+    #     while len(queue) > 0:
+    #         curr_content, curr_questions = queue.pop(0)
 
-            input_tokens_used = self.client.models.count_tokens(
-                model=self.model, contents = [system_prompt, curr_content, curr_questions]
-            )
+    #         input_tokens_used = self.client.models.count_tokens(
+    #             model=self.model, contents = [system_prompt, curr_content, curr_questions]
+    #         )
 
-            # Checking if the content is too large for the input token limit, if so splitting the content in half
-            # TODO: Add ability to use sliding window
-            if input_tokens_used > input_token_limit:
-                chunked_content = [curr_content[0: len(curr_content)//2 + 1], curr_content[len(curr_content)//2 + 1 : len(curr_content)]]
+    #         # Checking if the content is too large for the input token limit, if so splitting the content in half
+    #         # TODO: Add ability to use sliding window
+    #         if input_tokens_used > input_token_limit:
+    #             chunked_content = [curr_content[0: len(curr_content)//2 + 1], curr_content[len(curr_content)//2 + 1 : len(curr_content)]]
 
-                queue.append((chunked_content[0], curr_questions))
-                queue.append((chunked_content[1], curr_questions))
+    #             queue.append((chunked_content[0], curr_questions))
+    #             queue.append((chunked_content[1], curr_questions))
 
-            else:
-                query_contents = f'Content:\n{curr_content}\n\nThere are {len(curr_questions)} questions. The questions are:\n' + '\n\t- '.join(curr_questions)
-                response = self.generate_content(query_contents, system_prompt=system_prompt)
+    #         else:
+    #             query_contents = f'Content:\n{curr_content}\n\nThere are {len(curr_questions)} questions. The questions are:\n' + '\n\t- '.join(curr_questions)
+    #             response = self.generate_content(query_contents, system_prompt=system_prompt)
 
-                # TODO: This doesn't seem to actually occur, need a better way of doing this, checking if the output limit has been reached
-                if response.output_tokens > output_token_limit:
-                    batched_questions = TextChunkAndBatch.batch_by_number_of_questions(curr_questions, len(curr_questions)//2 + 1)
-                    queue.append((curr_content, batched_questions[0]))
-                    queue.append((curr_content, batched_questions[1]))
-                else:
-                    for i in range(len(response.content)):
-                        if curr_questions[i] not in answers.keys() and response.content[i] !=  'N/A':
-                            answers[curr_questions[i]] = response.content[i]
-                    total_input_tokens += response.input_tokens
-                    total_output_tokens += response.output_tokens
+    #             # TODO: This doesn't seem to actually occur, need a better way of doing this, checking if the output limit has been reached
+    #             if response.output_tokens > output_token_limit:
+    #                 batched_questions = TextChunkAndBatch.batch_by_number_of_questions(curr_questions, len(curr_questions)//2 + 1)
+    #                 queue.append((curr_content, batched_questions[0]))
+    #                 queue.append((curr_content, batched_questions[1]))
+    #             else:
+    #                 for i in range(len(response.content)):
+    #                     if curr_questions[i] not in answers.keys() and response.content[i] !=  'N/A':
+    #                         answers[curr_questions[i]] = response.content[i]
+    #                 total_input_tokens += response.input_tokens
+    #                 total_output_tokens += response.output_tokens
 
-        return Response(
-            content = answers,
-            input_tokens = total_input_tokens,
-            output_tokens = total_output_tokens
-        )
+    #     return Response(
+    #         content = answers,
+    #         input_tokens = total_input_tokens,
+    #         output_tokens = total_output_tokens
+    #     )
 
-    def generate_content_semantic(
-        self,
-        content : BaseTextInput,
-        questions : list[str],
-        system_prompt : str = None
-    ):
-        content_chunks, question_batches = TextChunkAndBatch.chunk_and_batch_semantically(content, questions)
+    # def generate_content_semantic(
+    #     self,
+    #     content : BaseTextInput,
+    #     questions : list[str],
+    #     system_prompt : str = None
+    # ):
+    #     content_chunks, question_batches = TextChunkAndBatch.chunk_and_batch_semantically(content, questions)
 
-        total_input_tokens = 0
-        total_output_tokens = 0
-        answers = {}
+    #     total_input_tokens = 0
+    #     total_output_tokens = 0
+    #     answers = {}
 
-        # Adding default system prompt if one is not given.
-        if system_prompt == None:
-            system_prompt = DEFAULT_SYSTEM_PROMPT
+    #     # Adding default system prompt if one is not given.
+    #     if system_prompt == None:
+    #         system_prompt = DEFAULT_SYSTEM_PROMPT
         
-        for i in range(len(content_chunks)):
-            # If there are no questions in the current chunk's batch, then we don't need to query it.
-            if len(question_batches[i]) != 0:
-                query_contents = f'Content:\n{content_chunks[i]}\n\nThere are {len(question_batches[i])} questions. The questions are:\n' + '\n\t- '.join(question_batches[i])
-                response = self.generate_content(query_contents, system_prompt=system_prompt)
+    #     for i in range(len(content_chunks)):
+    #         # If there are no questions in the current chunk's batch, then we don't need to query it.
+    #         if len(question_batches[i]) != 0:
+    #             query_contents = f'Content:\n{content_chunks[i]}\n\nThere are {len(question_batches[i])} questions. The questions are:\n' + '\n\t- '.join(question_batches[i])
+    #             response = self.generate_content(query_contents, system_prompt=system_prompt)
 
-                total_input_tokens += response.input_tokens
-                total_output_tokens += response.output_tokens
+    #             total_input_tokens += response.input_tokens
+    #             total_output_tokens += response.output_tokens
                 
-                for j in range(len(response.content)):
-                    answers[question_batches[i][j]] = response.content[j]
+    #             for j in range(len(response.content)):
+    #                 answers[question_batches[i][j]] = response.content[j]
         
-        return Response(
-            content = answers,
-            input_tokens = total_input_tokens,
-            output_tokens = total_output_tokens
-        )
+    #     return Response(
+    #         content = answers,
+    #         input_tokens = total_input_tokens,
+    #         output_tokens = total_output_tokens
+    #     )
 
-    def generate_content_media(
-        self,
-        media_path : str,
-        questions : list[str],
-        chunk_duration : int = 100,
-        questions_per_batch : int = 50,
-        window_duration : int = 0,
-        system_prompt : str = None
-    ):
+    # def generate_content_media(
+    #     self,
+    #     media_path : str,
+    #     questions : list[str],
+    #     chunk_duration : int = 100,
+    #     questions_per_batch : int = 50,
+    #     window_duration : int = 0,
+    #     system_prompt : str = None
+    # ):
         
-        # Adding default system prompt if one is not given.
-        if system_prompt == None:
-            system_prompt = DEFAULT_SYSTEM_PROMPT
+    #     # Adding default system prompt if one is not given.
+    #     if system_prompt == None:
+    #         system_prompt = DEFAULT_SYSTEM_PROMPT
 
-        chunker = MediaChunker()
-        number_of_chunks = chunker.sliding_window_chunking_by_duration(media_path, chunk_duration, window_duration)
+    #     chunker = MediaChunker()
+    #     number_of_chunks = chunker.sliding_window_chunking_by_duration(media_path, chunk_duration, window_duration)
 
-        question_batches = TextChunkAndBatch.batch_by_number_of_questions(questions, questions_per_batch)
+    #     question_batches = TextChunkAndBatch.batch_by_number_of_questions(questions, questions_per_batch)
 
-        answers = {}
+    #     answers = {}
 
-        total_input_tokens = 0
-        total_output_tokens = 0
+    #     total_input_tokens = 0
+    #     total_output_tokens = 0
 
-        for batch in question_batches:
-            for i in range(number_of_chunks):
-                query_contents = f'Content:\nThis has been attached as a media file, named {media_path}\n\nThere are {len(batch)} questions. The questions are:\n' + '\n\t- '.join(batch)
+    #     for batch in question_batches:
+    #         for i in range(number_of_chunks):
+    #             query_contents = f'Content:\nThis has been attached as a media file, named {media_path}\n\nThere are {len(batch)} questions. The questions are:\n' + '\n\t- '.join(batch)
                 
-                file = self.client.files.upload(file=f'./temp_output/chunk_{i}.mp4')
+    #             file = self.client.files.upload(file=f'./temp_output/chunk_{i}.mp4')
                 
-                response = self.generate_content(query_contents, system_prompt=system_prompt, file=file)
+    #             response = self.generate_content(query_contents, system_prompt=system_prompt, file=file)
 
-                total_input_tokens += response.input_tokens
-                total_output_tokens += response.output_tokens
+    #             total_input_tokens += response.input_tokens
+    #             total_output_tokens += response.output_tokens
 
-                # TODO: If the question has already been answered in a previous chunk the new answer is disregarded, this can be
-                # further optimised so the question is not asked again.
-                for i in range(len(response.content)):
-                    if batch[i] not in answers.keys() and response.content[i] !=  'N/A':
-                        answers[batch[i]] = response.content[i]
+    #             # TODO: If the question has already been answered in a previous chunk the new answer is disregarded, this can be
+    #             # further optimised so the question is not asked again.
+    #             for i in range(len(response.content)):
+    #                 if batch[i] not in answers.keys() and response.content[i] !=  'N/A':
+    #                     answers[batch[i]] = response.content[i]
 
 
-        return Response(
-            content = answers,
-            input_tokens = total_input_tokens,
-            output_tokens = total_output_tokens
-        )
+    #     return Response(
+    #         content = answers,
+    #         input_tokens = total_input_tokens,
+    #         output_tokens = total_output_tokens
+    #     )
